Moves the Bundle to the Public Folder
This commit moves the bundle created by Webpack to the public folder.
It also moves the files from within src/ to within client/.

Updates main component name from App to CalendarApp.
