+++ client/index.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import CalendarApp from './components/calendarApp.jsx';
+
+ReactDOM.render(<CalendarApp />, document.getElementById('calendar-app'));
