+++ public/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Calendar Module</title>
+  </head>
+  <body>
+    <div id='calendar-app'></div>
+    <script type="text/javascript" src='./bundle.js'></script>
+  </body>
+</html>
