+++ webpack.config.js
@@ -1,8 +1,8 @@
 module.exports = {
-  entry: __dirname + '/client/src/index.jsx',
+  entry: __dirname + '/client/index.jsx',
   output: {
     filename: 'bundle.js',
-    path: __dirname + '/client/dist'
+    path: __dirname + '/public'
   },
   module: {
     rules: [
